% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naverRank.R
\name{naverRank}
\alias{naverRank}
\title{Real-time search words ranking return on Naver.}
\usage{
naverRank(searchURL = "https://www.naver.com")
}
\arguments{
\item{searchURL}{Naver Real-time search ranking query.}
}
\value{
tibble type data.frame.
}
\description{
Real-time search words ranking return on Naver.
}
\examples{
naverRank()
}
