% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naverRelation.R
\name{naverRelation}
\alias{naverRelation}
\title{First & Second relation search words return on Naver.}
\usage{
naverRelation(keyword, depth = 1,
  searchURL = "https://search.naver.com/search.naver?where=nexearch&sm=top_hty&fbm=1&ie=utf8&query=")
}
\arguments{
\item{keyword}{A character of search keyword.}

\item{depth}{A number of relation depth you want to see(only 1 or 2).}

\item{searchURL}{Naver search query.}
}
\value{
tibble type data.frame.
}
\description{
First & Second relation search words return on Naver.
}
\examples{
naverRelation("banana", depth = 2)
}
