% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naverRelation.R
\name{naverRelation1}
\alias{naverRelation1}
\title{First relation search words return on Naver}
\usage{
naverRelation1(x)
}
\arguments{
\item{x}{a character of search keyword}
}
\value{
a vector; character type
}
\description{
First relation search words return on Naver
}
\examples{
naverRelation1("korea")
naverRelation1("apple")
}
