% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naverRelation.R
\name{naverRelation2}
\alias{naverRelation2}
\title{First & Second relation search words return on Naver}
\usage{
naverRelation2(x)
}
\arguments{
\item{x}{a character of search keyword}
}
\value{
tibble type data.frame
}
\description{
First & Second relation search words return on Naver
}
\examples{
naverRelation2("banana")
}
