% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{networkPlot}
\alias{networkPlot}
\title{Network plot for preprocessed naverRelation2 object}
\usage{
networkPlot(x)
}
\description{
2차 연관검색어에 대한 네트워크 시각화 함수입니다.
}
\examples{
network_graph <- naverRelation2("한국") \%>\%
  count(R2, R1) \%>\%
  graph_from_data_frame \%>\%
  networkPlot
}
