% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkplot_3d.R
\name{networkplot_3d}
\alias{networkplot_3d}
\title{Network plot about preprocessed naverRelation2 result}
\usage{
networkplot_3d(x)
}
\arguments{
\item{x}{return to \code{naverRelation2()} object}
}
\description{
Network plot about preprocessed naverRelation2 result
}
\examples{
networkPlot(naverRelation2("korea"))
}
