% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlotSave.R
\name{nvrGraph}
\alias{nvrGraph}
\title{Network plot about First & Second relation search words return on Naver}
\usage{
nvrGraph(input)
}
\arguments{
\item{input}{a character; search keyword}
}
\description{
Network plot about First & Second relation search words return on Naver
}
\examples{
nvrGraph("keyword")
}
