% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nr.R
\name{plot.nr}
\alias{plot.nr}
\title{Network plot about preprocessed naverRelation result}
\usage{
\method{plot}{nr}(x)
}
\arguments{
\item{x}{return to \code{naverRelation()} object}
}
\description{
Network plot about preprocessed naverRelation result
}
\examples{
relation <- naverRelation(keyword = "korea", depth = 2)
plot(relation) # same as plot.nr(relation)
}
