% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveToDB.R
\name{saveHistory}
\alias{saveHistory}
\alias{loadHistory}
\title{연관 검색 히스토리와 결과를 저장 및 로드하는 함수}
\usage{
saveHistory(df)

loadHistory()
}
\arguments{
\item{df}{\code{naverRelation2()} 함수를 통해 반환된 데이터프레임 타입 결과를 입력합니다.}
}
\description{
네이버 연관 검색어로 조회하였을 때 조회 히스토리를 DB로 저장하고 로드하는 함수입니다.
}
\examples{
res <- naverRelation2("Keyword")
saveHistory(res)
loadHistory()
}
