% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naverRelation.R
\name{naverRelation1}
\alias{naverRelation1}
\title{Naver 1차 연관검색어 추출}
\usage{
naverRelation1(x)
}
\arguments{
\item{x}{키워드를 캐릭터 형식으로 입력합니다}
}
\description{
특정 키워드를 입력하여 실행하면 그 키워드를 기준으로 검색되는 네이버 연관검색어가 벡터형으로 반환됩니다.
}
\examples{
naverRelation1("한국")
naverRelation1("사과")
}
