% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{networkPlot}
\alias{networkPlot}
\title{Network plot about preprocessed naverRelation2 result}
\usage{
networkPlot(x)
}
\arguments{
\item{x}{return to \code{naverRelation2()} object}
}
\description{
Network plot about preprocessed naverRelation2 result
}
\examples{
networkPlot(naverRelation2("korea"))
}
