% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nr.R
\name{plot.nr}
\alias{plot.nr}
\title{Network plot about preprocessed naverRelation result}
\usage{
\method{plot}{nr}(x)
}
\arguments{
\item{x}{return to \code{naverRelation()} object}
}
\description{
Network plot about preprocessed naverRelation result
}
\examples{
nr1 <- naverRelation("korea", 1)
plot(nr1)

nr2 <- naverRelation("korea", 2)
plot(nr2)
}
