% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveToDB.R
\name{saveHistory}
\alias{saveHistory}
\alias{loadHistory}
\title{Save and Load relation search words to sqlite}
\usage{
saveHistory(x, path = "nvrHistory.sqlite")

loadHistory()
}
\arguments{
\item{x}{return to \code{naverRelation()} object}

\item{path}{save fileDB path}
}
\description{
Save and Load relation search words to sqlite
}
\examples{
relation <- naverRelation("Keyword", 2)
saveHistory(relation)
loadHistory()
}
